weird (read desc)
Test case 6 was being read out of order, so when I changed the regex for digit, it fixed the issue. However, previously it recognized 77 as invalid, now it breaks it down into two separate single 7's. And the previous sample program is still read out of order in some parts. I will have to continue to play around with the regex combo's